+++ src/components-examples/angular/form-field/form-field-group/form-field-group-example.ts
@@ -1,5 +1,5 @@
 import { JsonPipe } from '@angular/common';
-import { Component } from '@angular/core';
+import { Component, inject } from '@angular/core';
 import { FormBuilder, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { SbbButtonModule } from '@sbb-esta/angular/button';
 import { SbbDatepickerModule } from '@sbb-esta/angular/datepicker';
@@ -25,13 +25,13 @@ import { SbbInputModule } from '@sbb-esta/angular/input';
   ],
 })
 export class FormFieldGroupExample {
+  private _formBuilder = inject(FormBuilder);
+
   formGroup = this._formBuilder.group({
     name: ['', [Validators.required, Validators.minLength(3)]],
     date: [new Date(), Validators.required],
   });
   lastSubmission?: { name?: string | null; date?: Date | null };
-
-  constructor(private _formBuilder: FormBuilder) {}
 
   handleSubmit() {
     if (!this.formGroup.valid) {